moved methods to store file
